automation_process/requirement.txt
fastapi
uvicorn
sqlalchemy
pandas
uuid
opencv-python
pillow
python-multipart
traceback